Restrict pid limiter to PowerVu only - Default set to 64 pids
The pid limiting feature is now applied to PowerVu channels only. The default value is again 64 pids, what oscam used before. The available options are 1-8, and 64 of course. You should not change this option unless you have a problematic box (some sh4 boxes like alien 2) that crashes with PowerVu channels. Thanks to joeuser for the help.